Markus Zojer:
I was in the "mood" to create a beast, the B-2, consisting of 3D model,
FDM and some "stolen" cockpit panels. Some autopilot functions are
misused for stabilizing the bird if you want to prevent your controls
from overstress. Since it is a bit annoying not to see other aircraft in
multiplayer when the aircraft model is not installed, feel free to add
this plane to the flightgear project.